remove unneeded inner check for chatsTosync.length
