+++ src/index.ts
@@ -1,4 +1,5 @@
 import * as api from '@/api/llm';
+import * as databaseApi from '@/api/database';
 import * as model from '@/models/response/models';
 import * as chatRequest from '@/models/request/chat_request';
 import * as content from '@/models/request/content';
@@ -14,9 +15,12 @@ import * as enums from '@/models/enums';
 import * as generateImageRequest from '@/models/request/generate_image_request';
 import * as imageResponse from '@/models/response/image_response';
 import * as sqllite from '@/models/sqllite';
+import * as databaseConfiguration from '@/models/storage/configuration';
+import * as databaseDto from '@/models/storage/dto';
 
 export { 
     api,
+    databaseApi,
     model,
     chatRequest,
     content,
@@ -31,5 +35,7 @@ export {
     enums,
     generateImageRequest,
     imageResponse,
-    sqllite
+    sqllite,
+    databaseConfiguration,
+    databaseDto
 };
