+++ src/models/storage/configuration.ts
@@ -1,0 +1,13 @@
+type DatabaseConfiguration = PostgresConfiguration;
+
+interface PostgresConfiguration {
+    connectionString: string;
+    ssl: {
+        rejectUnauthorized: boolean;
+    };
+};
+
+export {
+    DatabaseConfiguration,
+    PostgresConfiguration
+};
